+++ src/Lib.hs
@@ -1,9 +1,0 @@
-module Lib
-  ( libF
-  )
-where
-
-import           Control.Monad.IO.Class
-
-libF :: MonadIO m => m ()
-libF = liftIO $ putStrLn "hello, world ss"