+++ Changelog.md
@@ -1,6 +1,11 @@
-# Change log for template project
+# Change log for dropbox-client 
 
 ## Version 0.0.0 
 
-import [template](https://github.com/jappeace/haskell-template-project).
++ Add suppport for
+  + list folder (see what is in a users folder)
+  + get temporary link (download file support)
+  + dropbox token (allow us to get access to a users dropbox)
++ Add basic encoding/decoding tests for datatypes from dropbox api docs
++ import [template](https://github.com/jappeace/haskell-template-project).